requirements.txt
streamlit>=1.40.0
langchain==0.3.7
langchain-openai==0.2.0
langchain-community==0.3.0
chromadb>=0.4.0
langchain-chroma==0.1.4
pypdf>=5.0.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
